B: show favorited instance groups in dashboard
note: the prior behaviour when individual instances were favourited was
to display the instance group card in the favourites section with only
the favourited instances inside of it.

e.g. if instance group A has 10 instances, and 1 of those instances is
favourited, we'd show an instance group card for A with only 1 rectangle
inside

now, if instance group A is favourited and 1 of the instances is
favourited, we show an instance group card with all 10 rectangles
inside. the 1 instance being favourited currently has no effect (except
when drilling into instance groups)

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>