B: fix instance group card reordering bug
when the first pipeline of an instance group card is not visible (e.g.
because it is archived), the instance group card could not be
re-ordered. this was because we were using the id of the first pipeline
to determine uniqueness of cards, but when we re-order pipelines, we
start by collecting all the pipelines - not just the ones that are
visible.

this commit changes the identifier to a string - pipeline cards still
use the id of the pipeline as an identifier, while instance group cards
use "teamName/groupName" - this new identifier is independent of what
instances are visible

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>